some reorganization, csd move works
