+++ rosin/src/linux/csd_frame/frame.rs
@@ -5,7 +5,7 @@ use std::sync::Arc;
 use std::time::Duration;
 use std::{error::Error, num::NonZeroU32};
 
-use crate::linux::create_window::WaylandWindow;
+use crate::linux::wayland::WaylandWindow;
 use crate::linux::csd_frame::slot::SlotPool;
 use wayland_backend::client::ObjectId;
 use wayland_client::protocol::wl_compositor::WlCompositor;
